LPD-20754 Rename
