+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-etc-spring-boot/src/main/java/com/liferay/marketplace/AnalyticsRestController.java
@@ -36,7 +36,7 @@ public class AnalyticsRestController extends BaseRestController {
 		).get(
 		).uri(
 			uriBuilder -> uriBuilder.path(
-				"/reports/pages"
+				"/api/reports/pages"
 			).queryParam(
 				"channelId", channelId
 			).queryParam(
