+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-etc-spring-boot/src/main/resources/application-default.properties
@@ -4,7 +4,7 @@
 
 liferay.marketplace.analytics.auth.token=
 liferay.marketplace.analytics.auth.url=
-liferay.marketplace.console.auth.email=
+liferay.marketplace.console.auth.email.address=
 liferay.marketplace.console.auth.password=
 liferay.marketplace.console.auth.url=
 liferay.marketplace.console.cluster=
@@ -17,7 +17,7 @@ liferay.marketplace.koroneiki.auth.url=
 liferay.marketplace.provisioning.auth.client.id=
 liferay.marketplace.provisioning.auth.client.secret=
 liferay.marketplace.provisioning.auth.url=
-liferay.marketplace.trial.admin.email=
+liferay.marketplace.trial.admin.email.address=
 liferay.marketplace.trial.auth.client.id=
 liferay.marketplace.trial.auth.client.secret=
 liferay.marketplace.trial.auth.url=
